+++ R/themes.R
@@ -66,6 +66,7 @@ ggeffects_pal <- function(palette = "metro", n = NULL) {
 #' @export
 show_pals <- function() {
   insight::check_if_installed("ggplot2")
+  .data <- NULL
 
   longest.pal <- max(lengths(ggeffects_colors))
 
@@ -90,8 +91,8 @@ show_pals <- function() {
   x$group[x$key %in% c("breakfast.club", "flat", "metro", "quadro", "set1",
                        "simply", "social")] <- "Red-Blue-Green Palettes"
 
-  ggplot2::ggplot(x, ggplot2::aes_string(x = "key", fill = "cols")) +
-    ggplot2::geom_bar(width = .7) +
+  ggplot2::ggplot(x, ggplot2::aes(x = .data[["key"]], fill = .data[["cols"]])) +
+    ggplot2::geom_bar(width = 0.7) +
     ggplot2::scale_fill_manual(values = x$value) +
     ggplot2::scale_y_continuous(breaks = NULL, labels = NULL) +
     ggplot2::guides(fill = "none") +